backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
flask-cors
python-dotenv
shortuuid
python-slugify
moviepy
psycopg2-binary
gunicorn